🗃Refactor: 컴포넌트 template 분리
