+++ src/components/cafe/Header/Header.template.ts
@@ -1,17 +1,4 @@
-import { Component } from '@/components';
-
-export class Header extends Component {
-  bindEvents() {
-    this.$container.addEventListener('click', (event: MouseEvent) => {
-      const $target = event.target as HTMLInputElement;
-      if (!$target.matches('[type="radio"]')) return;
-      event.preventDefault();
-      alert('아직 준비되지 않았네요🥺');
-    });
-  }
-
-  template() {
-    return String.raw`
+export const template = String.raw`
 <header>
   <h1>☕ OOP 카페 ☕</h1>
   <nav>
@@ -23,5 +10,3 @@ export class Header extends Component {
   </nav>
 </header>
     `;
-  }
-}
