+++ src/components/cafe/Menu/Menu.template.ts
@@ -1,0 +1,34 @@
+export const template = String.raw`
+<div class='MenuSelect none-order'>
+  <h1>주방</h1>
+  <div class='coffee-container'>
+    <h1 class='coffee_name'>Choose your coffee</h1>
+    <div class='cup'>
+      <div class='filling reset'>
+        <div class='coffee'>커피</div>
+        <div class='water'>물</div>
+        <div class='liquor'>리퀴르</div>
+        <div class='milk'>우유</div>
+        <div class='whipped_cream'>휘핑 크림</div>
+        <div class='milk_foam'>밀크 폼</div>
+        <div class='steamed_milk'>데운 우유</div>
+        <div class='chocolate'>초콜릿</div>
+      </div>
+      <div class='plate'></div>
+    </div>
+  </div>
+  <div class='select-coffee-container'>
+    <div class='row'>
+      <div class='buttons'>
+      </div>
+    </div>
+    <div class='row'>
+      <div class='coffee-add-area'>
+        <form>
+          <button type='submit' class='coffee-add-options-button'>만들기</button>
+        </form>
+      </div>
+    </div>
+  </div>
+</div>    
+    `;
