+++ src/components/cafe/Order/OrderList.ts
@@ -2,6 +2,7 @@ import { EVENT } from '@/constant';
 import { Component, OrderListItem } from '@/components';
 import { createRandomOrder, Order } from '@/domain';
 import { dispatchCustomEvent } from '@/common';
+import { template } from './OrderList.template';
 
 export class OrderList extends Component {
   private $orderTable!: HTMLElement;
@@ -32,31 +33,6 @@ export class OrderList extends Component {
   }
 
   template() {
-    return String.raw`
-<div class='order-list'>
-  <h1>주문 목록</h1>
-  <div class='order-button-area'>
-    <button class='order-button'>주문 받기</button>
-  </div>
-  <div class='wrapper'>
-    <div class='table' id='order-table'>
-      <div class='table-row header'>
-        <div class='cell'>No</div>
-        <div class='cell'>메뉴명</div>
-        <div class='cell'>사이즈</div>
-        <div class='cell'>샷</div>
-        <div class='cell'>시럽</div>
-        <div class='cell'>ICE/HOT</div>
-        <div class='cell'>얼음 종류</div>
-        <div class='cell'>휘핑 크림</div>
-        <div class='cell'>엑스트라</div>
-        <div class='cell'>컵</div>
-        <div class='cell'>수정하기</div>
-        <div class='cell'>삭제하기</div>
-      </div>
-    </div>
-  </div>
-</div>    
-    `;
+    return template;
   }
 }
