+++ src/components/cafe/Serving/Serving.ts
@@ -1,0 +1,8 @@
+import { Component } from '@/components';
+import { template } from './Serving.template';
+
+export class Serving extends Component {
+  template() {
+    return template;
+  }
+}
