+++ src/components/index.ts
@@ -1,10 +1,10 @@
 export * from '@/components/core/Component';
 
-export * from '@/components/cafe/Header';
-export * from '@/components/cafe/Menu';
-export * from '@/components/cafe/Modal';
+export * from '@/components/cafe/Header/Header';
+export * from '@/components/cafe/Menu/Menu';
+export * from '@/components/cafe/Modal/Modal';
 
-export * from '@/components/cafe/OrderList';
-export * from '@/components/cafe/OrderListItem';
+export * from '@/components/cafe/Order/OrderList';
+export * from '@/components/cafe/Order/OrderListItem';
 
-export * from '@/components/cafe/Serving';
+export * from '@/components/cafe/Serving/Serving';
